Enable use with pre-17.4.0 versions of attr (#321)
* Enable use with pre-17.4.0 versions of attr.

PR #180 updated Grafanalib to work with attr versions post 19.2.0, when
keyword argument convert to attr.ib() was removed. It had been
deprecated in 17.4.0, and renamed to converter.

This change revises the change to work with both varieties of attr.ib(),
and so removes version requirements for attr. Those working on
antediluvian systems can use whatever system packaged version of attr is
available, a convenience when packaging Grafanalib.

* Set minimum supported attrs verison

Co-authored-by: Jim Hague <jim@sinodun.com>