Add missing format units (#494)
* Add boolean format units

These were added in Grafana 8.0.  They appear at the end of `categories.ts`, so
are added to the end here instead of alphabetized.

* Add missing Energy, Mass, and Length units

Current as of grafana eeaa160ae833.

* Add missing Miscellaneous units

Current as of grafana eeaa160ae833.

* Add missing Date & Time units

Current as of grafana eeaa160ae833.

* Add missing Data and Data Rate units

Current as of grafana eeaa160ae833.

* Fix typos in unit names

Client updates would be trivial, so I don't think it's worth keeping an alias to
the old names.